Add trader detail route and dialog; simplify dashboard handlers
